backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
google-generativeai>=0.5.2
PyMuPDF==1.24.2
reportlab==4.2.0
Pillow==10.3.0
python-dotenv==1.0.1
pydantic==2.7.0
pdfplumber==0.11.0
